refactor: streamline predict_instance_mask_from_image for improved FPS
- Remove unused dependencies
- Remove unused functions in utils.py and moved postprocess function to detector.py
- Consolidate methods in Detector() object
